Attempts to run NetSchool parser on the server (part 2)

+++ restart.sh
@@ -1,2 +1,7 @@
-kill `ps -ef | awk '/iterate.py/{print $2}' | head -n 1`
-python3 -u /home/panyuhin_nikita/NetSchool/iterate.py
+git fetch
+
+if [[ $(git rev-parse HEAD) != $(git rev-parse @{u}) ]]; then
+	kill `ps -ef | awk '/iterate.py/{print $2}' | head -n 1`
+	git pull --rebase
+	python3 -u /home/panyuhin_nikita/NetSchool/iterate.py
+fi
